Ops: automate the deployment of the Maple USDC pool collateral
